requirements.txt
langchain==0.1.12
transformers>=4.37.0
torch>=2.2.0
accelerate>=0.27.0
pydantic==2.6.4
python-dotenv==1.0.1
typing-extensions==4.10.0
aiohttp==3.10.5
nest-asyncio==1.5.8
groq>=0.3.2 